Simplify converting rst files to pdf and epub
The previous three scripts to convert rst into pdf, epub and single epub
were substituted by a single script with a simple bash script menu:
[P] For pdf files
[S] For single epub
[E] One epub per file

Also an automatic downloader for images was added in case the img/
directory at the root of the project is empty. The script connects to
dropbox and downloads a zip file, which one step ahead is unzipped and
placed at the root of the project.